ksonnet utilities for testing.
* Create a new module ks_util for common utilities related to ksonnet.

* Move the function get_ksonnet_cmd from run_e2e_workflow.py into ks_util

  * We want to reuse this function e.g. to submit TFJobs as part of E2E
    testing.

* Move setup_ks_app from
  https://github.com/kubeflow/tf-operator/blob/master/py/ks_util.py

  so it can be reused by other tests not just tf-operator.
